Improve naming of test classes
