+++ src/test/java/io/github/mirromutth/r2dbc/mysql/MySql55ConnectionIntegrationTest.java
@@ -17,11 +17,11 @@
 package io.github.mirromutth.r2dbc.mysql;
 
 /**
- * An implementation of {@link ConnectionTestSupport} for {@link MySqlConnection} unit tests in MySQL 8.0.
+ * An implementation of {@link ConnectionIntegrationTestSupport} for {@link MySqlConnection} unit tests in MySQL 5.5.
  */
-final class MySql80ConnectionTest extends ConnectionTestSupport {
+final class MySql55ConnectionIntegrationTest extends ConnectionIntegrationTestSupport {
 
-    MySql80ConnectionTest() {
-        super(MySql80TestKit.CONFIGURATION);
+    MySql55ConnectionIntegrationTest() {
+        super(MySql55TestKit.CONFIGURATION);
     }
 }
